renderer_vulkan: Move device abstraction to vulkan_common
